+++ agent/suite_test.go
@@ -2,11 +2,9 @@ package agent
 
 import (
 	"fmt"
-	"io/ioutil"
 	"log" // restrictpkg:ignore to suppress mysql client logs.
 	"os"
 	"path"
-	"strconv"
 	"testing"
 	"time"
 
@@ -32,41 +30,6 @@ const (
 )
 
 var replicationSourceSecretPath string
-
-type testData struct {
-	primaryHost            string
-	primaryPort            int
-	primaryUser            string
-	primaryPassword        string
-	cloneUser              string
-	clonePassword          string
-	initAfterCloneUser     string
-	initAfterClonePassword string
-}
-
-func writeTestData(data *testData) {
-	writeFile := func(filename, data string) error {
-		return ioutil.WriteFile(path.Join(replicationSourceSecretPath, filename), []byte(data), 0664)
-	}
-
-	var err error
-	err = writeFile("PRIMARY_HOST", data.primaryHost)
-	Expect(err).ShouldNot(HaveOccurred())
-	err = writeFile("PRIMARY_PORT", strconv.Itoa(data.primaryPort))
-	Expect(err).ShouldNot(HaveOccurred())
-	err = writeFile("PRIMARY_USER", data.primaryUser)
-	Expect(err).ShouldNot(HaveOccurred())
-	err = writeFile("PRIMARY_PASSWORD", data.primaryPassword)
-	Expect(err).ShouldNot(HaveOccurred())
-	err = writeFile("CLONE_USER", data.cloneUser)
-	Expect(err).ShouldNot(HaveOccurred())
-	err = writeFile("CLONE_PASSWORD", data.clonePassword)
-	Expect(err).ShouldNot(HaveOccurred())
-	err = writeFile("INIT_AFTER_CLONE_USER", data.initAfterCloneUser)
-	Expect(err).ShouldNot(HaveOccurred())
-	err = writeFile("INIT_AFTER_CLONE_PASSWORD", data.initAfterClonePassword)
-	Expect(err).ShouldNot(HaveOccurred())
-}
 
 func TestAgent(t *testing.T) {
 	mysql.SetLogger(mysql.Logger(log.New(GinkgoWriter, "[mysql] ", log.Ldate|log.Ltime|log.Lshortfile)))
